refactor: formatFileSize 함수 위치 변경에 따른 import 수정
- formatFileSize를 storage.ts에서 file-utils.ts로 이동
- FileUpload 컴포넌트의 import 경로 수정
